% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flora_version.R
\name{flora_version}
\alias{flora_version}
\title{Summarize Available Plant Taxonomic and Distributional Records from Flora e Funga do Brasil}
\usage{
flora_version()
}
\value{
A `data.frame` with the following columns:
\itemize{
  \item \code{Version}: Version number of the dataset release.
  \item \code{Published_on}: Timestamp of the dataset release.
  \item \code{Records}: Number of plant specimen records in the release.
  \item \code{URL}: Direct link to the dataset version on the IPT portal.
}
}
\description{
Retrieves and summarizes metadata of the currently available versions of plant
taxonomic and distributional records from the \href{https://floradobrasil.jbrj.gov.br/consulta/}{Flora e Funga do Brasil (FFB)},
hosted by the \href{https://www.gov.br/jbrj/pt-br}{Rio de Janeiro Botanical Garden}.
The metadata includes version number, publication date, number of records, and
direct access URLs.
}
\details{
This function accesses the \href{https://ipt.jbrj.gov.br/jbrj/resource?r=lista_especies_flora_brasil}{IPT metadata for the FFB dataset}
and extracts version information for each published data release. It returns
a cleaned and structured summary as a data frame.
}
\examples{
\dontrun{
  ffb_version <- flora_version()
  head(ffb_version)
}

}
\seealso{
\code{\link{flora_download}} for downloading the data of a specific version.
}
\author{
Domingos Cardoso & Kelmer Martins-Cunha
}
