% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floraR-package.R
\docType{package}
\name{floraR-package}
\alias{floraR}
\alias{floraR-package}
\title{floraR: Tools for Accessing, Analyzing, and Curating Data from the Flora e Funga do Brasil Platform}
\description{
The 'floraR' package provides a comprehensive interface to the Flora e Funga do Brasil (FFB) Platform, maintained by the Rio de Janeiro Botanical Garden. It offers tools for downloading, parsing, and analyzing taxonomic and distributional data published in Darwin Core Archive (DwC-A) format. Beyond data retrieval, 'floraR' also assists taxonomic experts contributing to the FFB by simplifying the integration of new species names and records from global biodiversity repositories such as IPNI, REFLORA, and GBIF. The package is designed to streamline both data exploration and data curation workflows, integrating seamlessly with the 'tidyverse' for downstream analyses and reproducible research pipelines.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dboslab.github.io/floraR-website/}
  \item \url{https://github.com/dboslab/floraR}
  \item Report bugs at \url{https://github.com/DBOSlab/floraR/issues}
}

}
\author{
\strong{Maintainer}: Domingos Cardoso \email{domingoscardoso@jbrj.gov.br} (\href{https://orcid.org/0000-0001-7072-2656}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Kelmer Martins-Cunha \email{kelmermartinscunha@gmail.com} (\href{https://orcid.org/0000-0002-0545-5966}{ORCID}) [copyright holder]
}

}
\keyword{internal}
